# packages/swagger/swagger.json
{
  "openapi": "3.1.0",
  "servers": [
    {
      "url": "https://github.com/samchon/nestia",
      "description": "insert your server url"
    }
  ],
  "info": {
    "version": "0.0.1",
    "title": "@IRIS/server",
    "description": "",
    "license": {
      "name": "AGPL-3.0-only"
    }
  },
  "components": {
    "schemas": {
...
      "IBadRequestException": {
        "additionalProperties": false,
        "description": "400 Bad Request exception payload.",
        "properties": {
          "statusCode": {
            "description": "HTTP status code corresponding to the error.",
            "readOnly": true,
            "type": "number"
          },
          "message": {
            "description": "Human-readable description of the error.",
            "readOnly": true,
            "type": "string"
          },
          "error": {
            "const": "Bad Request",
            "title": null,
            "description": "Standard error classification code.",
            "readOnly": true
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ],
        "type": "object"
      },
      "ISuccessResponse": {
        "additionalProperties": false,
        "description": "Standard operation status acknowledgment.",
        "properties": {
          "success": {
            "description": "Whether the operation succeeded.",
            "readOnly": true,
            "type": "boolean"
          },
          "message": {
            "description": "Optional status message.",
            "readOnly": true,
            "type": "string"
          }
        },
        "required": [
          "success",
          "message"
        ],
        "type": "object"
      },
...
      "IUnauthorizedException": {
        "additionalProperties": false,
        "description": "401 Unauthorized exception payload.",
        "properties": {
          "statusCode": {
            "description": "HTTP status code corresponding to the error.",
            "readOnly": true,
            "type": "number"
          },
          "message": {
            "description": "Human-readable description of the error.",
            "readOnly": true,
            "type": "string"
          },
          "error": {
            "const": "Unauthorized",
            "title": null,
            "description": "Standard error classification code.",
            "readOnly": true
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ],
        "type": "object"
      },
      "INotFoundException": {
        "additionalProperties": false,
        "description": "404 Not Found exception payload.",
        "properties": {
          "statusCode": {
            "description": "HTTP status code corresponding to the error.",
            "readOnly": true,
            "type": "number"
          },
          "message": {
            "description": "Human-readable description of the error.",
            "readOnly": true,
            "type": "string"
          },
          "error": {
            "const": "Not Found",
            "title": null,
            "description": "Standard error classification code.",
            "readOnly": true
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ],
        "type": "object"
      },
...
      "ILoginResponse": {
        "additionalProperties": false,
        "description": "Response payload returned upon successful primary login or MFA challenge.",
        "properties": {
          "success": {
            "description": "Whether authentication is fully complete.",
            "readOnly": true,
            "type": "boolean"
          },
          "mfaRequired": {
            "description": "True if secondary Multi-Factor Authentication is required to complete login.",
            "readOnly": true,
            "type": "boolean"
          },
          "mfaTicket": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ],
            "description": "Short-lived MFA challenge ticket when `mfaRequired` is true, or null.",
            "readOnly": true
          },
          "allowedMfaTypes": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "items": {
                  "oneOf": [
                    {
                      "const": "email"
                    },
                    {
                      "const": "passkey"
                    },
                    {
                      "const": "totp"
                    }
                  ]
                },
                "type": "array"
              }
            ],
            "description": "List of MFA verification methods available for the user.",
            "readOnly": true,
            "x-readonly-array": true
          },
          "user": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "additionalProperties": false,
                "properties": {
                  "id": {
                    "readOnly": true,
                    "type": "string"
                  },
                  "username": {
                    "readOnly": true,
                    "type": "string"
                  },
                  "email": {
                    "readOnly": true,
                    "type": "string"
                  },
                  "passwordChangedAt": {
                    "oneOf": [
                      {
                        "type": "null"
                      },
                      {
                        "type": "number"
                      }
                    ],
                    "readOnly": true
                  }
                },
                "required": [
                  "id",
                  "username",
                  "email",
                  "passwordChangedAt"
                ],
                "type": "object"
              }
            ],
            "description": "Authenticated user session summary or null if pending MFA.",
            "readOnly": true
          },
          "token": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ],
            "description": "JWT bearer session token or null if pending MFA.",
            "readOnly": true
          }
        },
        "required": [
          "success",
          "mfaRequired",
          "mfaTicket",
          "allowedMfaTypes",
          "user",
          "token"
        ],
        "type": "object"
      },
...
      "IPasskeyLoginOptionsResponse": {
        "additionalProperties": false,
        "description": "WebAuthn authentication options sent to the browser for `navigator.credentials.get()`.",
        "properties": {
          "challenge": {
            "description": "Cryptographic random challenge.",
            "readOnly": true,
            "type": "string"
          },
          "rpId": {
            "description": "Relying party identifier.",
            "readOnly": true,
            "type": "string"
          },
          "allowCredentials": {
            "description": "Allowed credentials for the user, if known.",
            "items": {
              "$ref": "#/components/schemas/IPublicKeyCredentialDescriptor"
            },
            "readOnly": true,
            "type": "array",
            "x-readonly-array": true
          },
          "timeout": {
            "description": "Challenge timeout in milliseconds.",
            "readOnly": true,
            "type": "number"
          },
          "userVerification": {
            "oneOf": [
              {
                "const": "discouraged"
              },
              {
                "const": "preferred"
              },
              {
                "const": "required"
              }
            ],
            "description": "User verification preference.",
            "readOnly": true
          }
        },
        "required": [
          "challenge",
          "rpId",
          "timeout",
          "userVerification"
        ],
        "type": "object"
      },
      "IPublicKeyCredentialDescriptor": {
        "additionalProperties": false,
        "description": "Public-key credential descriptor for WebAuthn allowed credentials.",
        "properties": {
          "id": {
            "description": "Base64URL-encoded credential ID.",
            "readOnly": true,
            "type": "string"
          },
          "type": {
            "const": "public-key",
            "title": null,
            "description": "Credential type (always 'public-key').",
            "readOnly": true
          },
          "transports": {
            "description": "Supported authenticator transports.",
            "items": {
              "type": "string"
            },
            "readOnly": true,
            "type": "array",
            "x-readonly-array": true
          }
        },
        "required": [
          "id",
          "type"
        ],
        "type": "object"
      },
      "IPasskeyLoginOptionsRequest": {
        "additionalProperties": false,
        "description": "Request payload for generating WebAuthn authentication options.",
        "properties": {
          "identifier": {
            "description": "Optional username or email to narrow down registered credentials on the client.",
            "minLength": 3,
            "readOnly": true,
            "type": "string"
          }
        },
        "required": [],
        "type": "object"
      },
      "IPasskeyVerifyLoginRequest": {
        "additionalProperties": false,
        "description": "Request payload for verifying a passkey assertion.",
        "properties": {
          "passkeyResponse": {
            "additionalProperties": false,
            "description": "WebAuthn assertion response object returned from `navigator.credentials.get()`.",
            "properties": {
              "id": {
                "readOnly": true,
                "type": "string"
              },
              "rawId": {
                "readOnly": true,
                "type": "string"
              },
              "response": {
                "additionalProperties": false,
                "properties": {
                  "clientDataJSON": {
                    "readOnly": true,
                    "type": "string"
                  },
                  "authenticatorData": {
                    "readOnly": true,
                    "type": "string"
                  },
                  "signature": {
                    "readOnly": true,
                    "type": "string"
                  },
                  "userHandle": {
                    "oneOf": [
                      {
                        "type": "null"
                      },
                      {
                        "type": "string"
                      }
                    ],
                    "readOnly": true
                  }
                },
                "readOnly": true,
                "required": [
                  "clientDataJSON",
                  "authenticatorData",
                  "signature"
                ],
                "type": "object"
              },
              "type": {
                "const": "public-key",
                "title": null,
                "description": null,
                "readOnly": true
              },
              "clientExtensionResults": {
                "$ref": "#/components/schemas/Recordstringstringnumberbooleanobject",
                "readOnly": true
              }
            },
            "readOnly": true,
            "required": [
              "id",
              "rawId",
              "response",
              "type"
            ],
            "type": "object"
          },
          "mfaTicket": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ],
            "description": "Optional MFA challenge ticket if completing secondary authentication.",
            "readOnly": true
          }
        },
        "required": [
          "passkeyResponse"
        ],
        "type": "object"
      },
      "Recordstringstringnumberbooleanobject": {
        "additionalProperties": {
          "oneOf": [
            {
              "type": "string"
            },
            {
              "type": "number"
            },
            {
              "type": "boolean"
            },
            {
              "$ref": "#/components/schemas/object"
            }
          ]
        },
        "type": "object"
      },
      "object": {
        "additionalProperties": false,
        "properties": {},
        "required": [],
        "type": "object"
      },
      "IPasskeyRegisterOptionsResponse": {
        "additionalProperties": false,
        "description": "WebAuthn registration options sent to the browser for `navigator.credentials.create()`.",
        "properties": {
          "challenge": {
            "description": "Cryptographic random challenge.",
            "readOnly": true,
            "type": "string"
          },
          "rp": {
            "additionalProperties": false,
            "description": "Relying party entity.",
            "properties": {
              "name": {
                "readOnly": true,
                "type": "string"
              },
              "id": {
                "readOnly": true,
                "type": "string"
              }
            },
            "readOnly": true,
            "required": [
              "name",
              "id"
            ],
            "type": "object"
          },
          "user": {
            "additionalProperties": false,
            "description": "User identity entity.",
            "properties": {
              "id": {
                "readOnly": true,
                "type": "string"
              },
              "name": {
                "readOnly": true,
                "type": "string"
              },
              "displayName": {
                "readOnly": true,
                "type": "string"
              }
            },
            "readOnly": true,
            "required": [
              "id",
              "name",
              "displayName"
            ],
            "type": "object"
          },
          "pubKeyCredParams": {
            "description": "Supported cryptographic algorithms.",
            "items": {
              "additionalProperties": false,
              "properties": {
                "alg": {
                  "readOnly": true,
                  "type": "number"
                },
                "type": {
                  "const": "public-key",
                  "title": null,
                  "description": null,
                  "readOnly": true
                }
              },
              "required": [
                "alg",
                "type"
              ],
              "type": "object"
            },
            "readOnly": true,
            "type": "array",
            "x-readonly-array": true
          },
          "timeout": {
            "description": "Challenge timeout in milliseconds.",
            "readOnly": true,
            "type": "number"
          },
          "authenticatorSelection": {
            "additionalProperties": false,
            "description": "Authenticator selection criteria.",
            "properties": {
              "residentKey": {
                "oneOf": [
                  {
                    "const": "discouraged"
                  },
                  {
                    "const": "preferred"
                  },
                  {
                    "const": "required"
                  }
                ],
                "readOnly": true
              },
              "userVerification": {
                "oneOf": [
                  {
                    "const": "discouraged"
                  },
                  {
                    "const": "preferred"
                  },
                  {
                    "const": "required"
                  }
                ],
                "readOnly": true
              }
            },
            "readOnly": true,
            "required": [],
            "type": "object"
          }
        },
        "required": [
          "challenge",
          "rp",
          "user",
          "pubKeyCredParams",
          "timeout"
        ],
        "type": "object"
      },
      "IPasskeyRegisterVerifyRequest": {
        "additionalProperties": false,
        "description": "Request payload for completing passkey registration.",
        "properties": {
          "passkeyResponse": {
            "additionalProperties": false,
            "description": "WebAuthn attestation response returned from `navigator.credentials.create()`.",
            "properties": {
              "id": {
                "readOnly": true,
                "type": "string"
              },
              "rawId": {
                "readOnly": true,
                "type": "string"
              },
              "response": {
                "additionalProperties": false,
                "properties": {
                  "clientDataJSON": {
                    "readOnly": true,
                    "type": "string"
                  },
                  "attestationObject": {
                    "readOnly": true,
                    "type": "string"
                  },
                  "transports": {
                    "items": {
                      "type": "string"
                    },
                    "readOnly": true,
                    "type": "array"
                  },
                  "publicKeyAlgorithm": {
                    "readOnly": true,
                    "type": "number"
                  },
                  "publicKey": {
                    "readOnly": true,
                    "type": "string"
                  },
                  "authenticatorData": {
                    "readOnly": true,
                    "type": "string"
                  }
                },
                "readOnly": true,
                "required": [
                  "clientDataJSON",
                  "attestationObject"
                ],
                "type": "object"
              },
              "type": {
                "const": "public-key",
                "title": null,
                "description": null,
                "readOnly": true
              },
              "clientExtensionResults": {
                "$ref": "#/components/schemas/Recordstringstringnumberbooleanobject",
                "readOnly": true
              }
            },
            "readOnly": true,
            "required": [
              "id",
              "rawId",
              "response",
              "type"
            ],
            "type": "object"
          },
          "name": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              }
            ],
            "description": "Optional custom label for the passkey (e.g. \"MacBook Touch ID\", \"YubiKey\").",
            "readOnly": true
          }
        },
        "required": [
          "passkeyResponse"
        ],
        "type": "object"
      },
...
    }
  },
  "paths": {
...
    "/auth/passkey/login-options": {
      "post": {
        "summary": "Generates WebAuthn authentication challenge options for passkey login",
        "description": "Generates WebAuthn authentication challenge options for passkey login.",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "description": "- Optional identifier.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IPasskeyLoginOptionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Challenge options for `navigator.credentials.get()`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IPasskeyLoginOptionsResponse"
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "auth",
          "passkey",
          "login_options",
          "getLoginOptions"
        ],
        "x-samchon-controller": "AuthPasskeyController"
      }
    },
    "/auth/passkey/verify-login": {
      "post": {
        "summary": "Verifies WebAuthn assertion signature and logs in the user",
        "description": "Verifies WebAuthn assertion signature and logs in the user.",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "description": "- Assertion response.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IPasskeyVerifyLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful login session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ILoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed clientDataJSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IBadRequestException"
                }
              }
            }
          },
          "401": {
            "description": "Passkey verification failed or challenge expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IUnauthorizedException"
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "auth",
          "passkey",
          "verify_login",
          "verifyLogin"
        ],
        "x-samchon-controller": "AuthPasskeyController"
      }
    },
    "/auth/passkey/register-options": {
      "post": {
        "summary": "Generates WebAuthn registration challenge options for an authenticated user",
        "description": "Generates WebAuthn registration challenge options for an authenticated user.",
        "tags": [],
        "parameters": [],
        "responses": {
          "201": {
            "description": "Registration options for `navigator.credentials.create()`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IPasskeyRegisterOptionsResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/INotFoundException"
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "auth",
          "passkey",
          "register_options",
          "getRegisterOptions"
        ],
        "x-samchon-controller": "AuthPasskeyController"
      }
    },
    "/auth/passkey/register-verify": {
      "post": {
        "summary": "Verifies and registers a new passkey",
        "description": "Verifies and registers a new passkey.",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "description": "- Attestation response and label.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IPasskeyRegisterVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Success acknowledgment.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ISuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Passkey registration verification failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IBadRequestException"
                }
              }
            }
          },
          "401": {
            "description": "Passkey registration challenge has expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IUnauthorizedException"
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "auth",
          "passkey",
          "register_verify",
          "verifyRegister"
        ],
        "x-samchon-controller": "AuthPasskeyController"
      }
    },
...
    "/auth/passkey/{id}": {
      "delete": {
        "summary": "Deletes a registered passkey",
        "description": "Deletes a registered passkey.",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "- ID of the passkey to delete."
          }
        ],
        "responses": {
          "200": {
            "description": "Success acknowledgment.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ISuccessResponse"
                }
              }
            }
          },
          "404": {
            "description": "Passkey not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/INotFoundException"
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "auth",
          "passkey",
          "deletePasskey"
        ],
        "x-samchon-controller": "AuthPasskeyController"
      }
    },
...
  },
...
}